d/aws_region: Remove terraform-plugin-sdk implementation.
