+++ internal/service/meta/region_data_source.go
@@ -1,101 +1,173 @@
 package meta
 
 import (
+	"context"
 	"fmt"
 	"strings"
 
 	"github.com/aws/aws-sdk-go/aws/endpoints"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
+	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 )
 
-func DataSourceRegion() *schema.Resource {
-	return &schema.Resource{
-		Read: dataSourceRegionRead,
+func init() {
+	registerFWDataSourceFactory(newDataSourceRegion)
+}
 
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Type:     schema.TypeString,
-				Optional: true,
-				Computed: true,
-			},
+// newDataSourceRegion instantiates a new DataSource for the aws_region data source.
+func newDataSourceRegion(context.Context) (datasource.DataSource, error) {
+	return &dataSourceRegion{}, nil
+}
 
-			"endpoint": {
-				Type:     schema.TypeString,
-				Optional: true,
-				Computed: true,
-			},
+type dataSourceRegion struct {
+	meta any
+}
 
+// Metadata should return the full name of the data source, such as
+// examplecloud_thing.
+func (d *dataSourceRegion) Metadata(_ context.Context, request datasource.MetadataRequest, response *datasource.MetadataResponse) {
+	response.TypeName = "aws_region"
+}
+
+// GetSchema returns the schema for this data source.
+func (d *dataSourceRegion) GetSchema(context.Context) (tfsdk.Schema, diag.Diagnostics) {
+	schema := tfsdk.Schema{
+		Attributes: map[string]tfsdk.Attribute{
 			"description": {
-				Type:     schema.TypeString,
+				Type:     types.StringType,
+				Computed: true,
+			},
+			"endpoint": {
+				Type:     types.StringType,
+				Optional: true,
+				Computed: true,
+			},
+			"id": {
+				Type:     types.StringType,
+				Optional: true,
+				Computed: true,
+			},
+			"name": {
+				Type:     types.StringType,
+				Optional: true,
 				Computed: true,
 			},
 		},
 	}
+
+	return schema, nil
 }
 
-func dataSourceRegionRead(d *schema.ResourceData, meta interface{}) error {
-	providerRegion := meta.(*conns.AWSClient).Region
+// Configure enables provider-level data or clients to be set in the
+// provider-defined DataSource type. It is separately executed for each
+// ReadDataSource RPC.
+func (d *dataSourceRegion) Configure(_ context.Context, request datasource.ConfigureRequest, response *datasource.ConfigureResponse) {
+	d.meta = request.ProviderData
+}
+
+// Read is called when the provider must read data source values in order to update state.
+// Config values should be read from the ReadRequest and new state values set on the ReadResponse.
+func (d *dataSourceRegion) Read(ctx context.Context, request datasource.ReadRequest, response *datasource.ReadResponse) {
+	var data dataSourceRegionData
+
+	response.Diagnostics.Append(request.Config.Get(ctx, &data)...)
+
+	if response.Diagnostics.HasError() {
+		return
+	}
 
 	var region *endpoints.Region
 
-	if v, ok := d.GetOk("endpoint"); ok {
-		endpoint := v.(string)
-		matchingRegion, err := FindRegionByEndpoint(endpoint)
+	if !data.Endpoint.IsNull() && !data.Endpoint.IsUnknown() {
+		matchingRegion, err := FindRegionByEndpoint(data.Endpoint.Value)
+
 		if err != nil {
-			return err
+			// TODO
+			response.Diagnostics.AddError("", "")
+
+			return
 		}
+
 		region = matchingRegion
 	}
 
-	if v, ok := d.GetOk("name"); ok {
-		name := v.(string)
-		matchingRegion, err := FindRegionByName(name)
+	if !data.Name.IsNull() && !data.Name.IsUnknown() {
+		matchingRegion, err := FindRegionByName(data.Name.Value)
+
 		if err != nil {
-			return err
+			// TODO
+			response.Diagnostics.AddError("", "")
+
+			return
 		}
+
 		if region != nil && region.ID() != matchingRegion.ID() {
-			return fmt.Errorf("multiple regions matched; use additional constraints to reduce matches to a single region")
+			// TODO
+			// "multiple regions matched; use additional constraints to reduce matches to a single region"
+			response.Diagnostics.AddError("", "")
+
+			return
 		}
+
 		region = matchingRegion
 	}
 
 	// Default to provider current region if no other filters matched
 	if region == nil {
-		matchingRegion, err := FindRegionByName(providerRegion)
+		matchingRegion, err := FindRegionByName(d.meta.(*conns.AWSClient).Region)
+
 		if err != nil {
-			return err
+			// TODO
+			response.Diagnostics.AddError("", "")
+
+			return
 		}
+
 		region = matchingRegion
 	}
 
-	d.SetId(region.ID())
+	regionEndpointEC2, err := region.ResolveEndpoint(endpoints.Ec2ServiceID)
 
-	regionEndpointEc2, err := region.ResolveEndpoint(endpoints.Ec2ServiceID)
 	if err != nil {
-		return err
+		// TODO
+		response.Diagnostics.AddError("", "")
+
+		return
 	}
-	d.Set("endpoint", strings.TrimPrefix(regionEndpointEc2.URL, "https://"))
 
-	d.Set("name", region.ID())
+	data.Description = types.String{Value: region.Description()}
+	data.Endpoint = types.String{Value: strings.TrimPrefix(regionEndpointEC2.URL, "https://")}
+	data.ID = types.String{Value: region.ID()}
+	data.Name = types.String{Value: region.ID()}
 
-	d.Set("description", region.Description())
+	response.Diagnostics.Append(response.State.Set(ctx, &data)...)
+}
 
-	return nil
+type dataSourceRegionData struct {
+	Description types.String `tfsdk:"description"`
+	Endpoint    types.String `tfsdk:"endpoint"`
+	ID          types.String `tfsdk:"id"`
+	Name        types.String `tfsdk:"name"`
 }
 
 func FindRegionByEndpoint(endpoint string) (*endpoints.Region, error) {
 	for _, partition := range endpoints.DefaultPartitions() {
 		for _, region := range partition.Regions() {
-			regionEndpointEc2, err := region.ResolveEndpoint(endpoints.Ec2ServiceID)
+			regionEndpointEC2, err := region.ResolveEndpoint(endpoints.Ec2ServiceID)
+
 			if err != nil {
 				return nil, err
 			}
-			if strings.TrimPrefix(regionEndpointEc2.URL, "https://") == endpoint {
+
+			if strings.TrimPrefix(regionEndpointEC2.URL, "https://") == endpoint {
 				return &region, nil
 			}
 		}
 	}
+
 	return nil, fmt.Errorf("region not found for endpoint %q", endpoint)
 }
 
@@ -107,5 +179,6 @@ func FindRegionByName(name string) (*endpoints.Region, error) {
 			}
 		}
 	}
+
 	return nil, fmt.Errorf("region not found for name %q", name)
 }
